Adds option to ignore toxic players
Extends the Toxicify addon with functionality to automatically add
toxic players to the ignore list. This provides users with a more
comprehensive way to manage interactions with players marked as toxic.

The changes include a new checkbox in the options panel to enable/disable
the automatic ignore feature, as well as updates to the marking logic
to add players to the ignore list when enabled.